+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ## Added
 
 - Add api endpoint to retrieve the list of course product relations
+- Add `max_validated_orders` field to CourseProductRelation model
 - Add `get_selling_organizations` method to Course model
 - Add courses client api endpoint
 - Add ThumbnailImageField "cover" to Course model